logic implementation

+++ src/main/java/com/amigoscode/demo/fee/Fee.java
@@ -1,11 +1,8 @@
 package com.amigoscode.demo.fee;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.math.BigDecimal;
 
 @Data
 @Document(collection = "fees")
@@ -13,6 +10,6 @@ public class Fee {
 
     @Id
     private String feeId;
-    private BigDecimal transactionValue;
+    private Double transactionValue;
     private double percentage;
 }
